glusterd: Use uint32_t for peerinfo generation number
Using a uint64_t for the peerinfo generation number was overkill for how
the generation number is used within GlusterD. It also prevented
GlusterD from running on 32-bit architechtures, as uatomic_add_return
doesn't support 64-bit values on 32-bit architechtures.

This change was developed on the git branch at [1]. This commit is a
combination of the following commits on the development branch.
  b78dba4 Use 32-bit generation number
  2c37e4b Change other generation number variables to uint32_t

[1]: https://github.com/kshlm/glusterfs/tree/urcu

BUG: 1205186
Reviewed-on: http://review.gluster.org/10425
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Tested-by: Emmanuel Dreyfus <manu@netbsd.org>
Tested-by: Krishnan Parthasarathi <kparthas@redhat.com>

+++ xlators/mgmt/glusterd/src/glusterd.h
@@ -172,7 +172,7 @@ typedef struct {
         gf_store_handle_t         *missed_snaps_list_shandle;
         struct cds_list_head       missed_snaps_list;
         int                        ping_timeout;
-        uint64_t                   generation;
+        uint32_t                   generation;
 } glusterd_conf_t;
